replace cairo-hints by scarb-agent
