+++ Cargo.toml
@@ -5,14 +5,14 @@ members = [
     "cairo-proto-serde",
     "cairo-proto-build",
     "cairo-oracle-hint-processor",
-    "cairo-hints",
+    "scarb-agent",
 ]
 exclude = ["examples/rust_complex/rust", "examples/rust_sqrt/rust"]
 
 [workspace.package]
 version = "0.1.1"
 edition = "2021"
-repository = "https://github.com/reilabs/cairo-hints/"
+repository = "https://github.com/gizatechxyz/cairo-hints/"
 license = "Apache-2.0"
 license-file = "LICENSE"
 
@@ -22,20 +22,20 @@ bincode = { version = "2.0.0-rc.3", default-features = false, features = [
     "serde",
 ] }
 bytes = "1.5.0"
-cairo-lang-casm = { version = "2.7.0" }
-cairo-lang-compiler = { version = "2.7.0" }
-cairo-lang-filesystem = { version = "2.7.0" }
-cairo-lang-runner = { version = "2.7.0" }
-cairo-lang-sierra = { version = "2.7.0" }
+cairo-lang-casm = "=2.7.0"
+cairo-lang-compiler = "=2.7.0"
+cairo-lang-filesystem = "=2.7.0"
+cairo-lang-runner = "=2.7.0"
+cairo-lang-sierra = "=2.7.0"
 cairo-lang-sierra-ap-change = { version = "2.7.0", default-features = false }
 cairo-lang-sierra-gas = { version = "2.7.0", default-features = false }
-cairo-lang-sierra-to-casm = { version = "2.7.0" }
-cairo-lang-sierra-generator = { version = "2.7.0" }
+cairo-lang-sierra-to-casm = "=2.7.0"
+cairo-lang-sierra-generator = "=2.7.0"
 cairo-lang-sierra-type-size = { version = "2.7.0", default-features = false }
-cairo-lang-starknet = { version = "2.7.0" }
+cairo-lang-starknet = "=2.7.0"
 cairo-lang-hints-test-runner = { path = "./cairo-lang-hints-test-runner", version = "0.1.0" }
-cairo-lang-test-plugin = { version = "2.7.0" }
-cairo-lang-utils = { version = "2.7.0" }
+cairo-lang-test-plugin = "=2.7.0"
+cairo-lang-utils = "=2.7.0"
 cairo-oracle-hint-processor = { path = "./cairo-oracle-hint-processor", version = "0.1.0" }
 cairo-proto-build = { path = "./cairo-proto-build", default-features = false, version = "0.1.0" }
 cairo-proto-serde = { path = "./cairo-proto-serde", version = "0.1.0" }
@@ -44,7 +44,7 @@ cairo-vm = { git = "https://github.com/lambdaclass/cairo-vm", rev = "v1.0.0", fe
     "std",
     "cairo-1-hints",
 ] }
-cainome-cairo-serde = {git = "https://github.com/cartridge-gg/cainome.git", rev = "0d29bb0"}
+cainome-cairo-serde = { git = "https://github.com/cartridge-gg/cainome.git", rev = "0d29bb0" }
 camino = "1.1.6"
 clap = { version = "4.3.10", features = ["derive"] }
 colored = "2.1.0"
