--- cairo-hints/scarb-hints-lib/lib.rs
@@ -1,2 +1,0 @@
-pub mod utils;
-pub mod serialization;
