+++ examples/rust_complex/Scarb.toml
@@ -8,7 +8,7 @@ edition = "2023_10"
 [dependencies]
 alexandria_math = { git = "https://github.com/keep-starknet-strange/alexandria.git", rev = "417c5a80ed2fef05ce159d643f3e34a5386ee06c" }
 
-[tool.hints]
+[tool.agent]
 definitions = "proto/oracle.proto" # must be provided
 # cairo_output = "src"                 # default "src"
 # oracle_lock = "Oracle.lock"          # default "Oracle.lock"
