+++ scarb-agent/scarb-agent-generate/main.rs
@@ -1,6 +1,6 @@
 use cairo_proto_build::Config;
 use clap::Parser;
-use scarb_hints_lib::utils::absolute_path;
+use scarb_agent_lib::utils::absolute_path;
 use scarb_metadata::MetadataCommand;
 use scarb_ui::args::PackagesFilter;
 use std::{io::Result, path::PathBuf};
@@ -32,24 +32,24 @@ fn main() -> Result<()> {
     let package = args.packages_filter.match_one(&metadata).unwrap();
 
     let definitions = absolute_path(&package, args.definitions, "definitions", None)
-        .expect("oracle.proto definitions path must be provided either as an argument (--definitions proto/oracle.proto) or in the Scarb.toml file in [tool.hints] section.");
+        .expect("oracle.proto definitions path must be provided either as an argument (--definitions proto/oracle.proto) or in the Scarb.toml file in [tool.agent] section.");
 
     let includes = definitions.parent().unwrap();
 
     let cairo_output: PathBuf = absolute_path(&package, args.cairo_output, "cairo_output", Some(PathBuf::from("src")))
-        .expect("cairo output path must be provided either as an argument (--cairo-output src) or in the Scarb.toml file in the [tool.hints] section.");
+        .expect("cairo output path must be provided either as an argument (--cairo-output src) or in the Scarb.toml file in the [tool.agent] section.");
 
     let oracle_module = args
         .oracle_module
         .or_else(|| {
             package
-                .tool_metadata("hints")
+                .tool_metadata("agent")
                 .and_then(|tool_config| tool_config["oracle_module"].as_str().map(String::from))
         })
         .unwrap_or("lib.cairo".to_string());
 
     let lock_output = absolute_path(&package, args.oracle_lock, "oracle_lock", Some(PathBuf::from("Oracle.lock")))
-        .expect("lock path must be provided either as an argument (--oracle-lock src) or in the Scarb.toml file in the [tool.hints] section.");
+        .expect("lock path must be provided either as an argument (--oracle-lock src) or in the Scarb.toml file in the [tool.agent] section.");
 
     Config::new()
         .out_dir(cairo_output)
