+++ scarb-agent/scarb-agent-lib/lib.rs
@@ -1,0 +1,2 @@
+pub mod serialization;
+pub mod utils;
