+++ scarb-agent/scarb-agent-lib/serialization.rs
