+++ scarb-agent/scarb-agent-new/new_python.rs
@@ -1,10 +1,10 @@
 use crate::fsx;
+use crate::templates::get_template_engine;
 use anyhow::Result;
 use camino::Utf8PathBuf;
 use once_cell::sync::Lazy;
 use scarb::core::{Config, PackageName};
 use serde_json::json;
-use crate::templates::get_template_engine;
 
 pub const SERVER_MANIFEST_PATH: Lazy<Utf8PathBuf> =
     Lazy::new(|| ["python", "requirements.txt"].iter().collect());
@@ -13,10 +13,10 @@ pub const SERVER_SOURCE_PATH: Lazy<Utf8PathBuf> =
 pub const GITIGNORE_PATH: Lazy<Utf8PathBuf> = Lazy::new(|| [".gitignore"].iter().collect());
 pub const README_PATH: Lazy<Utf8PathBuf> = Lazy::new(|| ["README.md"].iter().collect());
 pub const DOCKERFILE_PATH: Lazy<Utf8PathBuf> = Lazy::new(|| ["Dockerfile"].iter().collect());
-pub const PRE_COMMIT_CONFIG: Lazy<Utf8PathBuf> = Lazy::new(|| [".pre-commit-config.yaml"].iter().collect());
+pub const PRE_COMMIT_CONFIG: Lazy<Utf8PathBuf> =
+    Lazy::new(|| [".pre-commit-config.yaml"].iter().collect());
 pub const CLOUDBUILD_PATH: Lazy<Utf8PathBuf> = Lazy::new(|| ["cloudbuild.yaml"].iter().collect());
 pub const RUN_SERVICE_PATH: Lazy<Utf8PathBuf> = Lazy::new(|| ["run-service.yaml"].iter().collect());
-
 
 pub fn mk_python(canonical_path: &Utf8PathBuf, _: &PackageName, _config: &Config) -> Result<()> {
     // Get the templates registry
@@ -27,10 +27,7 @@ pub fn mk_python(canonical_path: &Utf8PathBuf, _: &PackageName, _config: &Config
     if !filename.exists() {
         fsx::create_dir_all(filename.parent().unwrap())?;
 
-        fsx::write(
-            filename,
-            registry.render("requirements", &json!({}))?,
-        )?;
+        fsx::write(filename, registry.render("requirements", &json!({}))?)?;
     }
 
     // Create the `Dockerfile` file.
@@ -38,10 +35,7 @@ pub fn mk_python(canonical_path: &Utf8PathBuf, _: &PackageName, _config: &Config
     if !filename.exists() {
         fsx::create_dir_all(filename.parent().unwrap())?;
 
-        fsx::write(
-            filename,
-            registry.render("dockerfile", &json!({}))?,
-        )?;
+        fsx::write(filename, registry.render("dockerfile", &json!({}))?)?;
     }
 
     // Create the `pre-commit` file.
@@ -50,10 +44,7 @@ pub fn mk_python(canonical_path: &Utf8PathBuf, _: &PackageName, _config: &Config
     if !filename.exists() {
         fsx::create_dir_all(filename.parent().unwrap())?;
 
-        fsx::write(
-            filename,
-            pre_commit,
-        )?;
+        fsx::write(filename, pre_commit)?;
     }
 
     // Create the `README.md` file.
@@ -61,20 +52,14 @@ pub fn mk_python(canonical_path: &Utf8PathBuf, _: &PackageName, _config: &Config
     let pre_commit = registry.render("readme", &json!({}))?;
     fsx::create_dir_all(filename.parent().unwrap())?;
 
-    fsx::write(
-        filename,
-        pre_commit,
-    )?;
+    fsx::write(filename, pre_commit)?;
 
     // Create the `main.py` file.
     let filename = canonical_path.join(SERVER_SOURCE_PATH.as_path());
     if !filename.exists() {
         fsx::create_dir_all(filename.parent().unwrap())?;
 
-        fsx::write(
-            filename,
-            registry.render("main", &json!({}))?
-        )?;
+        fsx::write(filename, registry.render("main", &json!({}))?)?;
     }
 
     // Create the `.gitignore` file.
@@ -82,10 +67,7 @@ pub fn mk_python(canonical_path: &Utf8PathBuf, _: &PackageName, _config: &Config
     if !filename.exists() {
         fsx::create_dir_all(filename.parent().unwrap())?;
 
-        fsx::write(
-            filename,
-            registry.render("gitignore", &json!({}))?
-        )?;
+        fsx::write(filename, registry.render("gitignore", &json!({}))?)?;
     }
 
     // Create the `cloudbuild.yaml` file.
@@ -93,10 +75,7 @@ pub fn mk_python(canonical_path: &Utf8PathBuf, _: &PackageName, _config: &Config
     if !filename.exists() {
         fsx::create_dir_all(filename.parent().unwrap())?;
 
-        fsx::write(
-            filename,
-            registry.render("cloudbuild", &json!({}))?
-        )?;
+        fsx::write(filename, registry.render("cloudbuild", &json!({}))?)?;
     }
 
     // Create the `run-service.yaml` file.
@@ -104,10 +83,7 @@ pub fn mk_python(canonical_path: &Utf8PathBuf, _: &PackageName, _config: &Config
     if !filename.exists() {
         fsx::create_dir_all(filename.parent().unwrap())?;
 
-        fsx::write(
-            filename,
-            registry.render("run-service", &json!({}))?
-        )?;
+        fsx::write(filename, registry.render("run-service", &json!({}))?)?;
     }
 
     Ok(())
