+++ scarb-agent/scarb-agent-new/new_rust.rs
