+++ scarb-agent/scarb-agent-new/restricted_names.rs
