+++ scarb-agent/scarb-agent-new/templates/python/.gitignore.hbs
