+++ scarb-agent/scarb-agent-new/templates/python/.pre-commit-config.hbs
