+++ scarb-agent/scarb-agent-new/templates/python/Dockerfile.hbs
