+++ scarb-agent/scarb-agent-new/templates/python/cloudbuild.hbs
