+++ scarb-agent/scarb-agent-new/templates/python/main.hbs
