+++ scarb-agent/scarb-agent-new/templates/python/requirements.hbs
