+++ scarb-agent/scarb-agent-new/templates/python/run-service.hbs
